Fixing SJ1 silkscreen
From Bug 8552: SJ1 had been labeled using the smash command, with the
name moved from tNames to tPlace.  Every board on the panel incremented
it's value (SJ1, SJ2 up to SJ15).

This makes the label on SJ1 a simple text object, which shouldn't
increment for each instance on the panel.